chore(components): close restructure debts — relocate samplrr, fix lint, sweep stale paths
- git mv components/samplrr -> components/platform/engines/samplrr (README-only
  slot, name unchanged per house style; no CMake wiring).
- Fix clang-diagnostic-sign-conversion in gui-sonotron capture_screenshot()
  (explicit size_t cast on height; behaviour identical).
- Sweep stale component-path mentions in code comments across apps/ and
  components/ (core/ + platform/ prefixes) and update retired
  gui_sonotron_audio::AudioEngine references to
  sonotron::audio::{SoundfontEngine,AudioBackend}.
- Docs prose (README, architecture, backlog-future, gui-and-ux,
  style-corpus) updated to the new core/platform paths; dated history
  (DESIGN.md §22, roadmap, reflections, proposals) left intact by design.

Known residual: a few samplrr doc mentions still read "not yet folded into
platform/" — deferred to a follow-up doc touch-up.

Host -Werror green, arm green, full ctest 126/126, 22 goldens byte-identical
(implementor + doc-steward reported gates; not re-run per owner directive).

+++ apps/gui-sonotron/main.cpp
@@ -154,7 +154,8 @@ void capture_screenshot(int width, int height, const char* path) {
   if (width <= 0 || height <= 0) {
     return;
   }
-  std::vector<unsigned char> pixels(static_cast<std::size_t>(width) * height * 4);
+  std::vector<unsigned char> pixels(static_cast<std::size_t>(width) *
+                                    static_cast<std::size_t>(height) * 4);
   glPixelStorei(GL_PACK_ALIGNMENT, 1);
   glReadPixels(0, 0, width, height, GL_RGBA, GL_UNSIGNED_BYTE, pixels.data());
   if (sonotron::write_png_rgba_bottom_up(path, width, height, pixels.data())) {
